requirements.txt
cffi==1.5.2
cryptography==1.3.1
decorator==4.0.6
Django==1.8.5
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-material==0.8.0
django-model-utils==2.5
django-notifications-hq==1.1
django-taggit==0.19.1
django-taggit-serializer==0.1.5
django-url-filter==0.2.0
django-utils==0.0.2
django-versatileimagefield==1.2.2
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
drf-nested-routers==0.11.1
drf-ujson==1.2
enum34==1.1.2
funcsigs==0.4
idna==2.1
ipaddress==1.0.16
ipython==4.0.3
ipython-genutils==0.1.0
jsonfield==1.0.3
ndg-httpsclient==0.4.0
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.0
psycopg2==2.6.1
ptyprocess==0.5
pusher==1.2.3
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==16.0.0
pytz==2016.4
requests==2.9.1
shortid==0.1.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.18
traitlets==4.1.0
ujson==1.35
urllib3==1.14
